Move scala.meta.cli utilities into their own packages
