+++ tests/jvm/src/test/scala/scala/meta/tests/api/SurfaceSuite.scala
@@ -44,8 +44,12 @@ class SurfaceSuite extends FunSuite {
       |scala.meta.classifiers.Classifiable *
       |scala.meta.classifiers.Classifier *
       |scala.meta.cli
-      |scala.meta.cli.Metac *
-      |scala.meta.cli.Metap *
+      |scala.meta.cli.metac *
+      |scala.meta.cli.metac.Main *
+      |scala.meta.cli.metacp *
+      |scala.meta.cli.metacp.Main *
+      |scala.meta.cli.metap *
+      |scala.meta.cli.metap.Main *
       |scala.meta.common
       |scala.meta.common.Convert *
       |scala.meta.common.Optional *
